Wrote API docs.
