+++ rust-postgres-rest-actix/README.md
@@ -129,6 +129,8 @@ Changing the previous API endpoint to `/api/child?columns=id,name,parent_id.name
 
 - Your tables & columns only contain letters, numbers, and underscore. We are converting query parameters/body parameters into an SQL abstract syntax tree (AST) before finally executing an SQL query in the background; there is no schema/model configuration (like in Diesel), so this restriction makes data easier to validate & secure.
 - You don’t need to query for `HStore`, `bit`, or `varbit` (technical limitations for now).
+- `actix-web >= 1.0.0`.
+- Rust >= 1.38 (that means nightly, for now).
 
 ## Configuration
 
@@ -557,6 +559,7 @@ In another terminal, run `cargo test`.
 ## To dos
 
 1. benchmark
+1. flatbuffers
 1. GraphQL API
 1. Optimization: Convert Strings to &str / statics.
 1. CSV, XML for REST API (nix for now?)
