+++ rust-postgres-rest-actix/src/endpoints/query_params_from_request.rs
@@ -1,9 +1,7 @@
 use crate::error::Error;
 use actix_web::HttpRequest;
 use rayon::prelude::*;
-use rust_postgres_rest::queries::query_types::{
-    DeleteParams, InsertParams, SelectParams, UpdateParams,
-};
+use rust_postgres_rest::queries::{DeleteParams, InsertParams, SelectParams, UpdateParams};
 use serde::Deserialize;
 use serde_json::{Map, Value};
