+++ rust-postgres-rest/README.md
@@ -1,3 +1,3 @@
 # rust-postgres-rest
 
-A module that contains the functions necessary to turn a PostgreSQL database into a REST API.
+A library that contains the functions necessary to turn a PostgreSQL database into a REST API.
