+++ rust-postgres-rest/src/queries/foreign_keys.rs
@@ -247,7 +247,7 @@ type OriginalColumnReferences = Vec<String>;
 
 #[derive(Debug)]
 /// Represents a single foreign key, usually generated by a queried column using dot-syntax.
-pub struct ForeignKeyReference {
+pub(crate) struct ForeignKeyReference {
     /// The original column strings referencing a (possibly nested) foreign key value.
     pub original_refs: OriginalColumnReferences,
